update 2026.05.27(1)
짧은 디버깅

+++ study-final-exam/prof'sExampleCode/2-my/src/Circle.java
@@ -15,7 +15,7 @@ public class Circle {
     public String toString(){
         return ("radius " + radius + " at " + p);
     }
-    public long calArea(){
-        return (long)(radius*radius*PI);
+    public double calArea(){
+        return (radius*radius*PI);
     }
 }
